# storage/api-docs/channel-operator-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Channel Operator API",
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Application root"
        }
    ],
    "paths": {
        "/api/v1/channels": {
            "get": {
                "tags": [
                    "Channels"
                ],
                "summary": "List channels visible to the authenticated user",
                "description": "Returns a paginated list of channels the authenticated user operates (channels they are a verified member of). Supports filtering by video-reception state, plus sorting and pagination. Admin-only fields such as weight and weekly quota are never exposed.",
                "operationId": "9bdb1228fc1d55f033e87acc04516814",
                "parameters": [
                    {
                        "name": "filter[is_video_reception_paused]",
                        "in": "query",
                        "description": "Exact match",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/SortNameCreatedAt"
                    },
                    {
                        "$ref": "#/components/parameters/PageNumber"
                    },
                    {
                        "$ref": "#/components/parameters/PageSize"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of channels",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Channel"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "channels:read"
                        ]
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/channels/{channel}": {
            "get": {
                "tags": [
                    "Channels"
                ],
                "summary": "Show a single channel",
                "description": "Returns one channel the user operates, by id. Responds with 404 if the channel does not exist or the user has no access to it.",
                "operationId": "5b893a62478f4df90cb8b727a29787e5",
                "parameters": [
                    {
                        "name": "channel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The channel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Channel"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "description": "Channel not found or not visible to the user"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "channels:read"
                        ]
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Channels"
                ],
                "summary": "Update channel settings",
                "description": "Changes the settings of a channel the user operates: the same settings the channel settings page offers. Every field is optional, send only what should change; fields that are left out keep their value. `name` and `email` cannot be cleared and must not be used by another channel, `creator_name` and `youtube_name` can be cleared with `null`. Setting `is_video_reception_paused` to `true` stops new offers for the channel, `show_on_homepage` controls whether the channel is listed on the public homepage. The logo has its own endpoints. Any other field, an empty body or a taken name or email is rejected with 422. Responds with 404 for channels not visible to the user.",
                "operationId": "35fa1d163213d971c753955720105498",
                "parameters": [
                    {
                        "name": "channel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "minProperties": 1,
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "creator_name": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "youtube_name": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 255
                                    },
                                    "is_video_reception_paused": {
                                        "type": "boolean"
                                    },
                                    "show_on_homepage": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated channel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Channel"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "description": "Channel not found or not visible to the user"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationFailed"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "channels:write"
                        ]
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/channels/{channel}/logo": {
            "post": {
                "tags": [
                    "Channels"
                ],
                "summary": "Upload a channel logo",
                "description": "Uploads a new logo for a channel the user operates and replaces the current one, which is deleted. Accepts PNG, JPEG or WebP up to 512 KB as multipart form field `logo`. Larger images are shrunk to fit into 256 x 256 pixels, keeping their aspect ratio and format; smaller images are stored as sent. The response contains the new `logo_url`. Responds with 422 for other file types, files over 512 KB or images that cannot be read, and with 404 for channels not visible to the user.",
                "operationId": "2d91e79b2b4b5d0f885e4bd919440bd2",
                "parameters": [
                    {
                        "name": "channel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "logo"
                                ],
                                "properties": {
                                    "logo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The channel with its new logo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Channel"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "description": "Channel not found or not visible to the user"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationFailed"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "channels:write"
                        ]
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Channels"
                ],
                "summary": "Remove a channel logo",
                "description": "Removes the logo of a channel the user operates and deletes the file. The channel is then shown without a logo and `logo_url` is `null`. Calling it for a channel without a logo succeeds as well. Responds with 404 for channels not visible to the user.",
                "operationId": "f04bd3ae412a3813ba29ec4d8b861444",
                "parameters": [
                    {
                        "name": "channel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The channel without logo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Channel"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "description": "Channel not found or not visible to the user"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "channels:write"
                        ]
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Channel": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "creator_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "youtube_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_video_reception_paused": {
                        "type": "boolean"
                    },
                    "show_on_homepage": {
                        "type": "boolean"
                    },
                    "logo_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "pagination": {
                        "properties": {
                            "current_page": {
                                "type": "integer"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "total": {
                                "type": "integer"
                            },
                            "total_pages": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "Forbidden": {
                "description": "Missing scope or permission"
            },
...
            "Unauthorized": {
                "description": "Unauthenticated"
            },
            "ValidationFailed": {
                "description": "Validation failed",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            }
        },
        "parameters": {
            "PageNumber": {
                "name": "page[number]",
                "in": "query",
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
            "PageSize": {
                "name": "page[size]",
                "in": "query",
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
            "SortNameCreatedAt": {
                "name": "sort",
                "in": "query",
                "description": "Comma-separated sort fields; prefix with \"-\" for descending. Allowed: name, created_at. Default: -created_at",
                "schema": {
                    "type": "string"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Paste a raw access token: a Personal Access Token created in the Standard panel, or a token obtained through the device grant.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "oauth2": {
                "type": "oauth2",
                "description": "OAuth2 access tokens. How to obtain one for interactive testing:\n\n- **authorization code**: run the flow below (the browser is redirected to\n  `/oauth/authorize`, the code is exchanged at `/oauth/token`). This produces a\n  user-scoped token and is the one to use for trying the endpoints out.\n- **personal access token**: create one in the Standard panel\n  (\"OAuth\" → \"Personal Access Tokens\") and paste it into the `bearerAuth` scheme.\n  This is the fastest path and needs no client setup.\n- **device**: request a code at `/oauth/device/code`, confirm it at `/oauth/device`,\n  poll `/oauth/token`, then paste the token into `bearerAuth`.\n- **client credentials**: also supported, but these tokens have no user context, so\n  every endpoint here returns 403 or an empty list. It is therefore not offered as an\n  interactive flow.\n\nScopes are `<resource>:<action>` and must be granted to the acting user.",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token",
                        "scopes": {
                            "channels:read": "List and read channels",
                            "channels:write": "Update channel settings",
                            "offers:read": "List and read offers",
                            "offers:download": "Download offered videos for accessible channels",
                            "offers:write": "Create offers and set comments"
                        }
                    }
                }
            }
        }
    },
...
}